dashboard/requirements.txt
# Investment Intelligence Dashboard Dependencies

# Web Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Visualization
plotly>=5.18.0

# Configuration
pyyaml>=6.0.1

# HTTP Requests (for potential API integrations)
requests>=2.31.0

# Date/Time Handling
python-dateutil>=2.8.2

# Caching
cachetools>=5.3.0

# Optional: For development
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.1.0
# mypy>=1.6.0
